+++ MonitorMirror/MonitorMirrorApp.swift
@@ -3,13 +3,11 @@ import SwiftUI
 @main
 struct MonitorMirrorApp: App {
     @StateObject private var peerSession = PeerSession()
-    @StateObject private var camera = CameraProcessor()
 
     var body: some Scene {
         WindowGroup {
             ContentView()
                 .environmentObject(peerSession)
-                .environmentObject(camera)
         }
     }
 }